rename staking modules, rename endpoints
